+++ src/childStoreAlt.ts
@@ -4,30 +4,28 @@ import { BaseEdgeStore } from './edgeStore'
 
 const _parentChildKey = Symbol('_parentChildKey')
 const _parentIdKey = Symbol('_parentIdKey')
+const _idKeyForChildRecord = Symbol('_idKeyForChildRecord')
+
+// This type allows us to strongly type a record with an un-conflictable property
+// that can store the parent ID, so that it can be serialized by toDb()
+type ChildRecord<Child> = Child & { [_idKeyForChildRecord]: string }
 
 export interface ChildStoreConfig<Parent, Child>
   extends Omit<StoreConfig<Child>, 'idKey' | 'sortKey'> {
-  /** property on the primary node that contains the id. */
+  /** property on the primary node that contains the id key. */
   parentIdKey: keyof Parent & string
   /** property on the primary node that contains an array of edges */
   parentChildKey: keyof Parent & string
-  /** property on the child node that contains its id. Used as the record's sortKey */
+  /** property on the child node that contains its id key. Used as the record's sortKey */
   childIdKey: keyof Child & string
-  /** property on the child node that contains its parent id key. Necessary to allow serialization of idKey: parentId, sortKey: childId */
-  childParentIdKey: keyof Child & string
 }
 
-export type ChildStoreSubConfig<Parent, Child> = Omit<
-  ChildStoreConfig<Parent, Child>,
-  'type' | 'parentIdKey' | 'parentChildKey' | 'childIdKey' | 'childParentIdKey'
->
-
-export class ChildStore<Parent, Child> extends BaseEdgeStore<Child> {
+export class ChildStore<Parent, Child> extends BaseEdgeStore<ChildRecord<Child>> {
   public readonly [_parentIdKey]: string
   public readonly [_parentChildKey]: string
 
   constructor(props: ChildStoreConfig<Parent, Child>) {
-    super({ ...props, idKey: props.childParentIdKey, sortKey: props.childIdKey })
+    super({ ...props, idKey: _idKeyForChildRecord, sortKey: props.childIdKey })
     if (!this[_dynamo][_sortField]) {
       throw new Error('ChildStore requires the dynamo client to have a sortField')
     }
@@ -61,16 +59,16 @@ export class ChildStore<Parent, Child> extends BaseEdgeStore<Child> {
     const dbEdges = await this.getByParentId(parentId)
 
     // Mapping with the parentId allows syncEdges to create the batchWrite with the correct idKey
-    return this.syncEdges(dbEdges, edges)
+    return this.syncEdges(dbEdges.map(withParentId(parentId)), edges.map(withParentId(parentId)))
   }
 
   /** Extract the children from the primary parent using the `_parentChildKey` defined in the constructor. Can be overridden */
   protected selectChildren(parent: Parent): Child[] {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    return ((parent as any)?.[this[_parentChildKey]] as Child[]) ?? []
+    return ((parent as any)?.[_parentChildKey] as Child[]) ?? []
   }
 
-  async getByParentId(parentId: string): Promise<Child[]> {
+  async getByParentId(parentId: string): Promise<ChildRecord<Child>[]> {
     return this.queryAll({
       ScanIndexForward: false,
       KeyConditionExpression: '#id = :id',
@@ -78,4 +76,8 @@ export class ChildStore<Parent, Child> extends BaseEdgeStore<Child> {
       ExpressionAttributeValues: { ':id': this.typeKey(parentId) },
     })
   }
+}
+
+function withParentId<Child>(parentId: string): (child: Child) => ChildRecord<Child> {
+  return (item) => ({ ...item, [_idKeyForChildRecord]: parentId })
 }