+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.2.0] - Unreleased
+## Added
+- `EdgeStore` for handling relationship records
+- `BaseEdgeStore` for handling relationship records manually
+
 ## [2.1.0] - 2021-10-18
 ## Added
 - `store.type`
@@ -17,7 +22,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `putAll`
 - `deleteAll`
 - `queryByPage`
-- `EdgeStore` class
 ## Removed
 - **BREAKING CHANGE** removed default value for `idKey` on `Store` constructor
 ## Deprecated